requirements.txt
fastapi
uvicorn[standard]
slowapi
prometheus-fastapi-instrumentator
pydantic
pydantic-settings
python-multipart
httpx
websockets
python-jose[cryptography]
qdrant-client
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
psycopg[binary]
docling
docling-core
beautifulsoup4
lxml
python-dateutil
pytz
tenacity

# Voice mode dependencies
numpy
faster-whisper
piper-tts

# Document parsing
pandas
pdfplumber
Pillow
pytesseract

# Media processing
ffmpeg-python

# Optional: install PyTorch by building Docker images with ENABLE_GPU_TORCH=1 when GPU acceleration is required.
